Extract API classes into their own artifact
 - Makes it easier to identify possible spec proposals in the future
 - Deprecated Emitter in annotations package, and created copy in main package as replacement
