+++ api/pom.xml
@@ -8,29 +8,13 @@
     <version>1.0.9-SNAPSHOT</version>
   </parent>
 
-  <artifactId>smallrye-reactive-messaging-provider</artifactId>
-
+  <artifactId>smallrye-reactive-messaging-api</artifactId>
+  <name>SmallRye Reactive Messaging API</name>
 
   <dependencies>
     <dependency>
-      <groupId>io.smallrye.reactive</groupId>
-      <artifactId>smallrye-reactive-messaging-api</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.apache.commons</groupId>
-      <artifactId>commons-lang3</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.jboss</groupId>
-      <artifactId>jandex</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>io.smallrye</groupId>
-      <artifactId>smallrye-config</artifactId>
-      <scope>test</scope>
+      <groupId>org.eclipse.microprofile.reactive.messaging</groupId>
+      <artifactId>microprofile-reactive-messaging-api</artifactId>
     </dependency>
   </dependencies>
