+++ api/src/main/java/io/smallrye/reactive/messaging/ChannelRegistar.java
