+++ api/src/main/java/io/smallrye/reactive/messaging/MediatorConfiguration.java
