+++ api/src/main/java/io/smallrye/reactive/messaging/Shape.java
