+++ api/src/main/java/io/smallrye/reactive/messaging/annotations/Merge.java
