+++ api/src/main/java/io/smallrye/reactive/messaging/annotations/Stream.java
@@ -12,10 +12,12 @@ import java.lang.annotation.Target;
 import javax.enterprise.util.Nonbinding;
 import javax.inject.Qualifier;
 
+import io.smallrye.reactive.messaging.Emitter;
+
 /**
  * This qualifier indicates which channel should be injected / populated.
  * <p>
- * This qualifier can be used to inject a <em>Channel</em> containing the items and signals propagated by the specified
+ * This qualifier can be used to inject a <em>stream</em> containing the items and signals propagated by the specified
  * channel. For example, it can be used to {@code @Inject} a {@code Publisher} representing a channel managed by the
  * Reactive Messaging implementation.
  * <p>
@@ -33,23 +35,25 @@ import javax.inject.Qualifier;
  *
  * <pre>
  * <code>
- * &#64;Inject @Channel("my-channel") Emitter&lt;String&gt; emitter;
+ * &#64;Inject @Stream("my-channel") Emitter&lt;String&gt; emitter;
  *
  * // ...
  * emitter.send("a").send("b").complete();
  * </code>
  * </pre>
  *
+ * @deprecated use {@link Channel @Channel} instead"
  */
 @Qualifier
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ METHOD, CONSTRUCTOR, FIELD, PARAMETER })
-public @interface Channel {
+@Deprecated
+public @interface Stream {
 
     /**
-     * The name of the channel (indicated in the {@code @Outgoing} annotation.
+     * The name of the stream (indicated in the {@code @Outgoing} annotation.
      *
-     * @return the channel name, mandatory, non null and non-blank. It must matches one of the available channels.
+     * @return the stream name, mandatory, non null and non-blank. It must matches one of the available streams.
      */
     @Nonbinding
     String value();
