+++ doc/advanced.adoc
@@ -72,7 +72,7 @@ Emitter<String> emitter;
 emitter.send("a").send("b");
 ----
 
-You can inject an `io.smallrye.reactive.messaging.annotations.Emitter<T>` or `io.smallrye.reactive.messaging.annotations.Emitter<Message<T>>`.
+You can inject an `io.smallrye.reactive.messaging.Emitter<T>` or `io.smallrye.reactive.messaging.Emitter<Message<T>>`.
 To use an `Emitter` for the channel `hello`, you need a `@Incoming("hello")` somewhere in your code (or in your configuration).
 
 When injecting an `Emitter` you can configure how the back-pressure is handled.
