+++ examples/amqp-quickstart/src/main/java/acme/BeanUsingAnEmitter.java
@@ -7,8 +7,8 @@ import java.util.concurrent.atomic.AtomicInteger;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
+import io.smallrye.reactive.messaging.Emitter;
 import io.smallrye.reactive.messaging.annotations.Channel;
-import io.smallrye.reactive.messaging.annotations.Emitter;
 
 @ApplicationScoped
 public class BeanUsingAnEmitter {
