+++ examples/mqtt-server-quickstart/src/main/java/acme/BeanUsingAnEmitter.java
@@ -1,7 +1,7 @@
 package acme;
 
 import io.smallrye.reactive.messaging.annotations.Channel;
-import io.smallrye.reactive.messaging.annotations.Emitter;
+import io.smallrye.reactive.messaging.Emitter;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
