+++ pom.xml
@@ -56,14 +56,18 @@
 
   <properties>
     <vertx.version>3.8.3</vertx.version>
-    <slf4j-api.version>1.7.29</slf4j-api.version>
-    <microprofile-reactive-messaging.version>1.0</microprofile-reactive-messaging.version>
-    <microprofile-reactive-streams.version>1.0.1</microprofile-reactive-streams.version>
     <rxjava.version>2.2.14</rxjava.version>
-    <smallrye-reactive-streams-ops.version>1.0.10</smallrye-reactive-streams-ops.version>
-    <smallrye-config.version>1.4.1</smallrye-config.version>
+    <slf4j-api.version>1.7.29</slf4j-api.version>
     <cloudevent.version>0.2.1</cloudevent.version>
     <axle.version>0.0.10</axle.version>
+
+    <cdi.version>2.0.SP1</cdi.version>
+    <microprofile-reactive-messaging.version>1.0</microprofile-reactive-messaging.version>
+    <microprofile-reactive-streams.version>1.0.1</microprofile-reactive-streams.version>
+    <microprofile-config.version>1.3</microprofile-config.version>
+
+    <smallrye-reactive-streams-ops.version>1.0.10</smallrye-reactive-streams-ops.version>
+    <smallrye-config.version>1.4.1</smallrye-config.version>
 
     <asciidoctorj.version>1.5.8.1</asciidoctorj.version>
 
@@ -71,6 +75,7 @@
   </properties>
 
   <modules>
+    <module>api</module>
     <module>smallrye-reactive-messaging-provider</module>
     <module>smallrye-reactive-messaging-kafka</module>
     <module>smallrye-reactive-messaging-mqtt</module>
@@ -118,41 +123,67 @@
       </dependency>
 
       <dependency>
+        <groupId>org.eclipse.microprofile.reactive.messaging</groupId>
+        <artifactId>microprofile-reactive-messaging-api</artifactId>
+        <version>${microprofile-reactive-messaging.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.microprofile.reactive-streams-operators</groupId>
+        <artifactId>microprofile-reactive-streams-operators-api</artifactId>
+        <version>${microprofile-reactive-streams.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.microprofile.reactive-streams-operators</groupId>
+        <artifactId>microprofile-reactive-streams-operators-core</artifactId>
+        <version>${microprofile-reactive-streams.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.microprofile.config</groupId>
+        <artifactId>microprofile-config-api</artifactId>
+        <version>${microprofile-config.version}</version>
+      </dependency>
+
+      <dependency>
         <groupId>io.smallrye</groupId>
         <artifactId>smallrye-config</artifactId>
         <version>${smallrye-config.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>org.apache.commons</groupId>
+        <artifactId>commons-lang3</artifactId>
+        <version>3.9</version>
+      </dependency>
+
+      <dependency>
+        <groupId>io.reactivex.rxjava2</groupId>
+        <artifactId>rxjava</artifactId>
+        <version>${rxjava.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>javax.enterprise</groupId>
+        <artifactId>cdi-api</artifactId>
+        <version>${cdi.version}</version>
+        <scope>provided</scope>
+      </dependency>
+
+      <!-- From this project -->
+      <dependency>
+        <groupId>io.smallrye.reactive</groupId>
+        <artifactId>smallrye-reactive-messaging-api</artifactId>
+        <version>${project.version}</version>
       </dependency>
     </dependencies>
   </dependencyManagement>
 
   <dependencies>
+    <!-- logging -->
     <dependency>
-      <groupId>org.eclipse.microprofile.reactive.messaging</groupId>
-      <artifactId>microprofile-reactive-messaging-api</artifactId>
-      <version>${microprofile-reactive-messaging.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.microprofile.reactive-streams-operators</groupId>
-      <artifactId>microprofile-reactive-streams-operators-api</artifactId>
-      <version>${microprofile-reactive-streams.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.microprofile.reactive-streams-operators</groupId>
-      <artifactId>microprofile-reactive-streams-operators-core</artifactId>
-      <version>${microprofile-reactive-streams.version}</version>
+      <groupId>org.slf4j</groupId>
+      <artifactId>slf4j-api</artifactId>
     </dependency>
 
-
-    <dependency>
-      <groupId>io.reactivex.rxjava2</groupId>
-      <artifactId>rxjava</artifactId>
-      <version>${rxjava.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>javax.enterprise</groupId>
-      <artifactId>cdi-api</artifactId>
-      <scope>provided</scope>
-    </dependency>
     <dependency>
       <groupId>org.jboss.weld.se</groupId>
       <artifactId>weld-se-core</artifactId>
@@ -190,17 +221,6 @@
     </dependency>
 
     <dependency>
-      <groupId>org.eclipse.microprofile.config</groupId>
-      <artifactId>microprofile-config-api</artifactId>
-      <version>1.3</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.commons</groupId>
-      <artifactId>commons-lang3</artifactId>
-      <version>3.9</version>
-    </dependency>
-
-    <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <scope>test</scope>
@@ -214,12 +234,6 @@
       <groupId>org.awaitility</groupId>
       <artifactId>awaitility</artifactId>
       <scope>test</scope>
-    </dependency>
-
-    <!-- logging -->
-    <dependency>
-      <groupId>org.slf4j</groupId>
-      <artifactId>slf4j-api</artifactId>
     </dependency>
   </dependencies>
