+++ smallrye-reactive-messaging-provider/src/test/java/io/smallrye/reactive/messaging/inject/overflow/BufferOverflowStrategyTest.java
@@ -22,9 +22,9 @@ import io.reactivex.Flowable;
 import io.reactivex.Scheduler;
 import io.reactivex.exceptions.MissingBackpressureException;
 import io.reactivex.schedulers.Schedulers;
+import io.smallrye.reactive.messaging.Emitter;
 import io.smallrye.reactive.messaging.WeldTestBaseWithoutTails;
 import io.smallrye.reactive.messaging.annotations.Channel;
-import io.smallrye.reactive.messaging.annotations.Emitter;
 import io.smallrye.reactive.messaging.annotations.OnOverflow;
 
 public class BufferOverflowStrategyTest extends WeldTestBaseWithoutTails {
