+++ smallrye-reactive-messaging-provider/src/test/java/io/smallrye/reactive/messaging/inject/overflow/DropOverflowStrategyTest.java
@@ -21,9 +21,9 @@ import org.junit.Test;
 import io.reactivex.Flowable;
 import io.reactivex.Scheduler;
 import io.reactivex.schedulers.Schedulers;
+import io.smallrye.reactive.messaging.Emitter;
 import io.smallrye.reactive.messaging.WeldTestBaseWithoutTails;
 import io.smallrye.reactive.messaging.annotations.Channel;
-import io.smallrye.reactive.messaging.annotations.Emitter;
 import io.smallrye.reactive.messaging.annotations.OnOverflow;
 
 public class DropOverflowStrategyTest extends WeldTestBaseWithoutTails {
